+++ src/constants/index.ts
@@ -1,2 +1,3 @@
 export * from './ui';
 export * from './training';
+export * from './units';